refactor(LottoRanks): 로또 당첨 정보를 관리하는 맵을 일급 컬렉션으로 변경
Statistics에서 모두 처리했던 당첨 금액과 수익률을 구하는 기능 분리

+++ src/main/java/lotto/view/OutputView.java
@@ -27,8 +27,7 @@ public class OutputView {
         }
     }
 
-
-    public static void printWinResult(Map<LottoRank, Integer> winLottoInfo) {
+    public static void printWinResult(Map<LottoRank, Long> winLottoInfo) {
         System.out.println(View.OUTPUT_STATISTICS.message());
         System.out.println(View.OUTPUT_DOTTED_LINE.message());
 
@@ -37,7 +36,7 @@ public class OutputView {
                 .filter(rank -> !rank.equals(LottoRank.FAIL))
                 .sorted(Comparator.comparingInt(LottoRank::matchCount))
                 .sorted((o1, o2) -> Boolean.compare(o1.isBonusNumber(), o2.isBonusNumber()))
-                .forEach(rank -> printResultAccordingToBonus(rank, winLottoInfo.get(rank)));
+                .forEach(rank -> printResultAccordingToBonus(rank, winLottoInfo.get(rank).intValue()));
     }
 
     private static void printResultAccordingToBonus(LottoRank rank, int totalCount) {
